+++ TrkPatRec/src/RobustHelixFinder_module.cc
@@ -93,24 +93,24 @@ namespace mu2e {
     int                                 _printfreq;
     bool				_prefilter; // prefilter hits based on sector
     bool				_updatestereo; // update the stereo hit positions each iteration
-    double				_dhit; // distance between hit position updates to consider changed' (mm)
-    double				_dhit2;
+    float				_dhit; // distance between hit position updates to consider changed' (mm)
+    float				_dhit2;
     unsigned				_minnhit; // minimum # of hits to work with
-    double				_maxdr; // maximum hit-helix radius difference
-    double				_maxrpull; // maximum hit-helix radius difference pull
-    double				_maxphisep; // maximum separation in global azimuth of hits
+    float				_maxdr; // maximum hit-helix radius difference
+    float				_maxrpull; // maximum hit-helix radius difference pull
+    float				_maxphisep; // maximum separation in global azimuth of hits
     TrkFitFlag				_saveflag; // write out all helices that satisfy these flags
     unsigned				_maxniter;  // maximum # of iterations over outlier filtering + fitting
-    double				_cradres; // average center resolution along center position (mm)
-    double				_cperpres; // average center resolution perp to center position (mm)
-    double				_maxdwire; // outlier cut on distance between hit and helix along wire
-    double				_maxdtrans; // outlier cut on distance between hit and helix perp to wire
-    double				_maxchisq; // outlier cut on chisquared
-    double				_maxrwdot[2]; // outlier cut on angle between radial direction and wire: smaller is better
-    double				_minrerr; // minimum radius error
+    float				_cradres; // average center resolution along center position (mm)
+    float				_cperpres; // average center resolution perp to center position (mm)
+    float				_maxdwire; // outlier cut on distance between hit and helix along wire
+    float				_maxdtrans; // outlier cut on distance between hit and helix perp to wire
+    float				_maxchisq; // outlier cut on chisquared
+    float				_maxrwdot[2]; // outlier cut on angle between radial direction and wire: smaller is better
+    float				_minrerr; // minimum radius error
 
     bool				_usemva; // use MVA to cut outliers
-    double				_minmva; // outlier cut on MVA
+    float				_minmva; // outlier cut on MVA
 
     art::InputTag			_ccTag;
     art::InputTag			_chTag;
@@ -126,7 +126,7 @@ namespace mu2e {
     RobustHelixFit   _hfit;
     std::vector<Helicity> _hels; // helicity values to fit 
     TrkTimeCalculator _ttcalc;
-    double            _t0shift;   
+    float            _t0shift;   
     StrawHitFlag      _outlier;
     bool              _updateStereo;
     
@@ -151,21 +151,21 @@ namespace mu2e {
     _printfreq   (pset.get<int>("printFrequency",101)),
     _prefilter   (pset.get<bool>("PrefilterHits",true)),
     _updatestereo(pset.get<bool>("UpdateStereoHits",true)),
-    _dhit	 (pset.get<double>("HitDistanceChange",10.0)), // mm
+    _dhit	 (pset.get<float>("HitDistanceChange",10.0)), // mm
     _minnhit	 (pset.get<unsigned>("minNHit",5)),
-    _maxdr	 (pset.get<double>("MaxRadiusDiff",100.0)), // mm
-    _maxrpull	 (pset.get<double>("MaxRPull",5.0)), // unitless
-    _maxphisep	 (pset.get<double>("MaxPhiHitSeparation",1.0)),
+    _maxdr	 (pset.get<float>("MaxRadiusDiff",100.0)), // mm
+    _maxrpull	 (pset.get<float>("MaxRPull",5.0)), // unitless
+    _maxphisep	 (pset.get<float>("MaxPhiHitSeparation",1.0)),
     _saveflag    (pset.get<vector<string> >("SaveHelixFlag",vector<string>{"HelixOK"})),
     _maxniter    (pset.get<unsigned>("MaxIterations",10)), // iterations over outlier removal
-    _cradres	 (pset.get<double>("CenterRadialResolution",20.0)),
-    _cperpres	 (pset.get<double>("CenterPerpResolution",12.0)),
-    _maxdwire    (pset.get<double>("MaxWireDistance",200.0)), // max distance along wire
-    _maxdtrans   (pset.get<double>("MaxTransDistance",80.0)), // max distance perp to wire (and z)
-    _maxchisq    (pset.get<double>("MaxChisquared",100.0)), // max chisquared
-    _minrerr     (pset.get<double>("MinRadiusErr",20.0)), // mm
+    _cradres	 (pset.get<float>("CenterRadialResolution",20.0)),
+    _cperpres	 (pset.get<float>("CenterPerpResolution",12.0)),
+    _maxdwire    (pset.get<float>("MaxWireDistance",200.0)), // max distance along wire
+    _maxdtrans   (pset.get<float>("MaxTransDistance",80.0)), // max distance perp to wire (and z)
+    _maxchisq    (pset.get<float>("MaxChisquared",100.0)), // max chisquared
+    _minrerr     (pset.get<float>("MinRadiusErr",20.0)), // mm
     _usemva      (pset.get<bool>("UseHitMVA",true)),
-    _minmva      (pset.get<double> ("MinMVA",0.1)), // min MVA output to define an outlier
+    _minmva      (pset.get<float> ("MinMVA",0.1)), // min MVA output to define an outlier
     _ccTag	 (pset.get<art::InputTag>("CaloClusterCollection","CaloClusterFast")),
     _chTag	 (pset.get<art::InputTag>("ComboHitCollection","makeSH")),
     _tcTag	 (pset.get<art::InputTag>("TimeClusterCollection","TimeClusterFinder")),
@@ -175,12 +175,12 @@ namespace mu2e {
     _nsmva       (pset.get<fhicl::ParameterSet>("HelixNonStereoHitMVA",fhicl::ParameterSet())),
     _hfit        (pset.get<fhicl::ParameterSet>("RobustHelixFit",fhicl::ParameterSet())),
     _ttcalc      (pset.get<fhicl::ParameterSet>("T0Calculator",fhicl::ParameterSet())),
-    _t0shift     (pset.get<double>("T0Shift",4.0)),
+    _t0shift     (pset.get<float>("T0Shift",4.0)),
     _outlier     (StrawHitFlag::outlier),
     _updateStereo    (pset.get<bool>("UpdateStereo",true))
   {
-    _maxrwdot[0] = pset.get<double>("MaxStereoRWDot",1.0);
-    _maxrwdot[1] = pset.get<double>("MaxNonStereoRWDot",1.0);
+    _maxrwdot[0] = pset.get<float>("MaxStereoRWDot",1.0);
+    _maxrwdot[1] = pset.get<float>("MaxNonStereoRWDot",1.0);
     _dhit2 = _dhit*_dhit;
     std::vector<int> helvals = pset.get<std::vector<int> >("Helicities",vector<int>{Helicity::neghel,Helicity::poshel});
     for(auto hv : helvals) {
@@ -288,10 +288,10 @@ namespace mu2e {
 
       _vmva._dtrans = fabs(dh.Dot(wtdir));              // transverse projection
       _vmva._dwire = fabs(dh.Dot(wdir));               // projection along wire direction
-      _vmva._drho = fabs(sqrt(cvec.mag2()) - helix.radius()); // radius difference
+      _vmva._drho = fabs(sqrtf(cvec.mag2()) - helix.radius()); // radius difference
       _vmva._dphi = fabs(hhit.helixPhi() - helix.circleAzimuth(hhit.pos().z())); // azimuth difference WRT circle center
-      _vmva._hhrho = sqrt(cvec.mag2());            // hit transverse radius WRT circle center
-      _vmva._hrho = sqrt(hpos.Perp2());            // hit detector transverse radius
+      _vmva._hhrho = sqrtf(cvec.mag2());            // hit transverse radius WRT circle center
+      _vmva._hrho = sqrtf(hpos.Perp2());            // hit detector transverse radius
       _vmva._rwdot = fabs(wdir.Dot(cdir));  // compare directions of radius and wire
 
       // compute the total resolution including hit and helix parameters first along the wire
@@ -304,7 +304,7 @@ namespace mu2e {
 	std::pow(_cradres*cdir.Dot(wtdir),(int)2) +
 	std::pow(_cperpres*cperp.Dot(wtdir),(int)2);
 
-      _vmva._chisq = sqrt( _vmva._dwire*_vmva._dwire/wres2 + _vmva._dtrans*_vmva._dtrans/wtres2 );          
+      _vmva._chisq = sqrtf( _vmva._dwire*_vmva._dwire/wres2 + _vmva._dtrans*_vmva._dtrans/wtres2 );          
       _vmva._dt = hhit.time() - hseed._t0.t0();
 
       if (hhit._flag.hasAnyProperty(StrawHitFlag::stereo))
@@ -347,11 +347,11 @@ namespace mu2e {
       float rwdot = wdir.Dot(cdir); // compare directions of radius and wire
       float rwdot2 = rwdot*rwdot;
       // compute radial difference and pull
-      float dr = sqrt(cvec.mag2())-helix.radius();
+      float dr = sqrtf(cvec.mag2())-helix.radius();
       float werr = hhit.posRes(StrawHitPosition::wire);
       float terr = hhit.posRes(StrawHitPosition::trans);
       // the resolution is dominated the resolution along the wire
-      float rres = std::max(sqrt(werr*werr*rwdot2 + terr*terr*(1.0-rwdot2)),_minrerr);
+      float rres = std::max(sqrtf(werr*werr*rwdot2 + terr*terr*(1.0-rwdot2)),_minrerr);
       float rpull = dr/rres;
       unsigned ist = hhit._flag.hasAnyProperty(StrawHitFlag::stereo) ? 0 : 1;
 
@@ -380,8 +380,8 @@ namespace mu2e {
     {     
       if (hhit._flag.hasAnyProperty(_outlier)) continue;
 
-      double hphi = hhit.pos().phi();
-      double dphi = fabs(Angles::deltaPhi(hphi,helix.fcent()));
+      float hphi = hhit.pos().phi();
+      float dphi = fabs(Angles::deltaPhi(hphi,helix.fcent()));
 
       const XYZVec& wdir = hhit.wdir();
       XYZVec wtdir = zaxis.Cross(wdir);   // transverse direction to the wire
@@ -392,19 +392,19 @@ namespace mu2e {
       XYZVec hpos = hhit.pos(); // this sets the z position to the hit z
       helix.position(hpos);                // this computes the helix expectation at that z
       XYZVec dh = hhit.pos() - hpos;   // this is the vector between them
-      double dtrans = fabs(dh.Dot(wtdir)); // transverse projection
-      double dwire = fabs(dh.Dot(wdir));   // projection along wire direction
+      float dtrans = fabs(dh.Dot(wtdir)); // transverse projection
+      float dwire = fabs(dh.Dot(wdir));   // projection along wire direction
 
       // compute the total resolution including hit and helix parameters first along the wire
-      double wres2 = std::pow(hhit.posRes(StrawHitPosition::wire),(int)2) +
+      float wres2 = std::pow(hhit.posRes(StrawHitPosition::wire),(int)2) +
 	std::pow(_cradres*cdir.Dot(wdir),(int)2) +
 	std::pow(_cperpres*cperp.Dot(wdir),(int)2);
       // transverse to the wires
-      double wtres2 = std::pow(hhit.posRes(StrawHitPosition::trans),(int)2) +
+      float wtres2 = std::pow(hhit.posRes(StrawHitPosition::trans),(int)2) +
 	std::pow(_cradres*cdir.Dot(wtdir),(int)2) +
 	std::pow(_cperpres*cperp.Dot(wtdir),(int)2);
 
-      double chisq = dwire*dwire/wres2 + dtrans*dtrans/wtres2;
+      float chisq = dwire*dwire/wres2 + dtrans*dtrans/wtres2;
 
       if( dphi > _maxphisep || fabs(dwire) > _maxdwire || fabs(dtrans) > _maxdtrans || chisq > _maxchisq) 
       {
@@ -425,8 +425,8 @@ namespace mu2e {
     {
       nhit = 0;
       changed = false;
-      accumulator_set<double, stats<tag::median(with_p_square_quantile) > > accx;
-      accumulator_set<double, stats<tag::median(with_p_square_quantile) > > accy;
+      accumulator_set<float, stats<tag::median(with_p_square_quantile) > > accx;
+      accumulator_set<float, stats<tag::median(with_p_square_quantile) > > accy;
       for (const auto& hhit : hhits ) 
       {
 	if (hhit._flag.hasAnyProperty(_outlier)) continue;
@@ -435,17 +435,17 @@ namespace mu2e {
 	++nhit;
       }
 
-      double mx = extract_result<tag::median>(accx);
-      double my = extract_result<tag::median>(accy);
-      double mphi = atan2(my,mx);
+      float mx = extract_result<tag::median>(accx);
+      float my = extract_result<tag::median>(accy);
+      float mphi = atan2f(my,mx);
 
-      double maxdphi{0.0};
+      float maxdphi{0.0};
       auto worsthit = hhits.end();
       for(auto ihit = hhits.begin(); ihit != hhits.end(); ++ihit)
       {
 	if (ihit->_flag.hasAnyProperty(_outlier)) continue;
-	double phi  = ihit->pos().phi();
-	double dphi = fabs(Angles::deltaPhi(phi,mphi));
+	float phi  = ihit->pos().phi();
+	float dphi = fabs(Angles::deltaPhi(phi,mphi));
 	if(dphi > maxdphi)
 	{
 	  maxdphi = dphi;
@@ -464,25 +464,25 @@ namespace mu2e {
   void RobustHelixFinder::updateT0(HelixSeed& hseed)
   {
     const auto& hhits = hseed.hits();
-    accumulator_set<double, stats<tag::weighted_variance(lazy)>, double > terr;
+    accumulator_set<float, stats<tag::weighted_variance(lazy)>, float > terr;
     for (const auto& hhit : hhits) 
     {
       if (hhit._flag.hasAnyProperty(_outlier)) continue;
-      double wt = std::pow(1.0/_ttcalc.strawHitTimeErr(),2);
+      float wt = std::pow(1.0/_ttcalc.strawHitTimeErr(),2);
       terr(_ttcalc.comboHitTime(hhit),weight=wt);
     }
 
     if (hseed.caloCluster().isNonnull())
     {
-      double time = _ttcalc.caloClusterTime(*hseed.caloCluster());
-      double wt = std::pow(1.0/_ttcalc.caloClusterTimeErr(hseed.caloCluster()->diskId()),2);
+      float time = _ttcalc.caloClusterTime(*hseed.caloCluster());
+      float wt = std::pow(1.0/_ttcalc.caloClusterTimeErr(hseed.caloCluster()->diskId()),2);
       terr(time,weight=wt);
     }
 
     if (sum_of_weights(terr) > 0.0)
     {
       hseed._t0._t0 = extract_result<tag::weighted_mean>(terr) + _t0shift; // ad-hoc correction FIXME!!
-      hseed._t0._t0err = sqrt(std::max(0.0,extract_result<tag::weighted_variance(lazy)>(terr))/extract_result<tag::count>(terr));
+      hseed._t0._t0err = sqrtf(std::max(float(0.0),extract_result<tag::weighted_variance(lazy)>(terr))/extract_result<tag::count>(terr));
     }
   }
