+++ TrkReco/inc/TLTClusterer.hh
@@ -38,10 +38,10 @@ namespace mu2e
   private:
   // functions to define the distance between a hit and cluster
   // this is the variable against which the clustering thresholds are applied
-  double distance(BkgCluster const&, ComboHit const& ch) const; 
+  float distance(BkgCluster const&, ComboHit const& ch) const; 
   // merge overlapping clusters
   unsigned mergeClusters(BkgClusterCollection& bkgcol,
-      double dt, double drho2,
+      float dt, float drho2,
       ComboHitCollection const& chcol ) const;
  // compress out empties
   void cleanClusters(BkgClusterCollection& bkgcol ) const;
@@ -51,13 +51,13 @@ namespace mu2e
   // are tested first, for efficiency
   void assignHits(BkgClusterCollection& bkgcol, ComboHitCollection const& chcol) const;
   // compute the total distance between hits and clusters
-  double hitDistance(BkgClusterCollection const& bkgcol, ComboHitCollection const& chcol) const;
+  float hitDistance(BkgClusterCollection const& bkgcol, ComboHitCollection const& chcol) const;
   // update the cluster state (position, distances)
   void updateClusters(BkgClusterCollection& clusters, ComboHitCollection const& chcol) const;
   //
   void updateCluster(BkgCluster& cluster, ComboHitCollection const& chcol) const;
   //
-  void countClusters(BkgClusterCollection const& clusters, unsigned& nclu, unsigned& nchits, double& tdist) const;
+  void countClusters(BkgClusterCollection const& clusters, unsigned& nclu, unsigned& nchits, float& tdist) const;
   void refineClusters(BkgClusterCollection& clusters, ComboHitCollection const& chcol) const;
   void refineCluster(BkgCluster& cluster, ComboHitCollection const& chcol ) const;
 
@@ -67,24 +67,24 @@ namespace mu2e
   StrawHitFlag _bkgmask; // mask for hits to ignore
   StrawHitFlag _sigmask; // mask for hits to use
   StrawHitFlag _stereo; // mask for selecting stereo hits
-  double _dseed; // Minimum separation to seed a new cluster
-  double _dhit; // Maximum separation to include a hit in a cluster
-  double _dd; // cluster diameter
-  double _dd2; // cluster diameter squared, cached for efficiency
-  double _dt; // natural time spread
-  double _maxdt; // maximum time difference
-  double _trms2; // time RMS
-  double _md2; // // cached square of maximum distance
-  double _maxdsum; // maximum total distance change to consider 'converged'
+  float _dseed; // Minimum separation to seed a new cluster
+  float _dhit; // Maximum separation to include a hit in a cluster
+  float _dd; // cluster diameter
+  float _dd2; // cluster diameter squared, cached for efficiency
+  float _dt; // natural time spread
+  float _maxdt; // maximum time difference
+  float _trms2; // time RMS
+  float _md2; // // cached square of maximum distance
+  float _maxdsum; // maximum total distance change to consider 'converged'
   unsigned _maxniter; // maximum number of iterations
   bool _stereoinit; // initialize using only stereo hits
-  double _maxwt; // maximum weight to give a single hit
-  double _minerr; // minimum error to give a single hit
+  float _maxwt; // maximum weight to give a single hit
+  float _minerr; // minimum error to give a single hit
 
   // diagnostics
   mutable TTree* _idiag; // iteration diagnostics
   mutable unsigned _niter, _nmerge, _nclu, _nhits, _nchits;
-  mutable double _odist, _tdist;
+  mutable float _odist, _tdist;
   };
 }
 #endif
