+++ TrkReco/inc/TNTBClusterer.hh
@@ -16,20 +16,20 @@ namespace mu2e {
 
    struct ClusterStrawHit2 
    {      
-       ClusterStrawHit2(size_t index, const ComboHit& ch, double psig2inv) : 
+       ClusterStrawHit2(size_t index, const ComboHit& ch, float psig2inv) : 
          _index(index),_dist(10000),_time(ch.time()),_itime(int(_time/10.0)),_pos(ch.pos()),_wdir(ch.wdir()),
          _phi(_pos.phi()),_psig2inv(psig2inv),_posResInv(1.0/ch.posRes(ComboHit::wire)),_nchanged(0)
        {}   
 
        size_t            _index;
-       double            _dist;
-       double            _time;
+       float            _dist;
+       float            _time;
        int               _itime;
        XYZVec _pos;
        XYZVec _wdir;
-       double            _phi;
-       double            _psig2inv;
-       double            _posResInv;
+       float            _phi;
+       float            _psig2inv;
+       float            _posResInv;
        unsigned          _nchanged;
    };
 
@@ -42,19 +42,19 @@ namespace mu2e {
         virtual ~ClusterStraw2() {};
 
         inline const  XYZVec& pos()              const { return _pos; }
-        inline double time()                                const { return _time; }
+        inline float time()                                const { return _time; }
         inline int    itime()                               const { return _itime; }
         inline bool   hasChanged()                          const { return _hasChanged;}
         inline void   flagChanged(bool val)                       { _hasChanged = val;}
         inline const  std::vector<ClusterStrawHit2*>& hits() const { return _hitsPtr; }
         inline        std::vector<ClusterStrawHit2*>& hits()       { return _hitsPtr; }
                 
-        void updateCache(double _maxwt);
+        void updateCache(float _maxwt);
   
   
      private:
         XYZVec             _pos;
-        double                        _time;
+        float                        _time;
         int                           _itime;
         std::vector<ClusterStrawHit2*> _hitsPtr; 
         bool                          _hasChanged;
@@ -76,25 +76,25 @@ namespace mu2e {
      private:
 
          unsigned formClusters(std::vector<ClusterStrawHit2>& chits, std::list<ClusterStraw2>& clusters, bool init);
-         double   distance(const ClusterStraw2&, ClusterStrawHit2&) const;
-         double   distance2(const ClusterStraw2&, ClusterStrawHit2&) const;
+         float   distance(const ClusterStraw2&, ClusterStrawHit2&) const;
+         float   distance2(const ClusterStraw2&, ClusterStrawHit2&) const;
          void     dump(std::list<ClusterStraw2>& clusters);
 
          int          _diag;
          StrawHitFlag _bkgmask; // mask for background hits
          StrawHitFlag _sigmask; // mask for selecting signals
          bool         _stereoInit;  // Start with stereo hits
-         double       _dseed;       // Minimum separation to seed a new cluster
-         double       _dhit;        // Maximum separation to include a hit in a cluster
-         double       _dmerge;      // distance to merge 2 clusters
-         double       _dd;          // cluster diameter
-         double       _dt;          // natural time spread
-         double       _maxdt;       // maximum time difference
-         double       _trms;        // time RMS
-         double       _maxdsum; 
-         double       _minerr; // Spatial RMS for stereo, non-stereo hits
-         double       _maxdist; // Maximum transverse distance (mm)
-         double       _trms2inv, _srms2inv, _nsrms2inv; // squares of rms
+         float       _dseed;       // Minimum separation to seed a new cluster
+         float       _dhit;        // Maximum separation to include a hit in a cluster
+         float       _dmerge;      // distance to merge 2 clusters
+         float       _dd;          // cluster diameter
+         float       _dt;          // natural time spread
+         float       _maxdt;       // maximum time difference
+         float       _trms;        // time RMS
+         float       _maxdsum; 
+         float       _minerr; // Spatial RMS for stereo, non-stereo hits
+         float       _maxdist; // Maximum transverse distance (mm)
+         float       _trms2inv, _srms2inv, _nsrms2inv; // squares of rms
          unsigned     _maxniter;    // maximum number of iterations
          unsigned     _maxnchanged;   
 
@@ -103,11 +103,11 @@ namespace mu2e {
 	 int        _ditime;
          unsigned   _niter;  
          unsigned   _nchanged; 
-         double     _md2;
-         double     _dd2; 
-         double     _maxwt; 
-         double     _tdist;
-         double     _odist;
+         float     _md2;
+         float     _dd2; 
+         float     _maxwt; 
+         float     _tdist;
+         float     _odist;
          int        _nclu;
          int        _nhits, _nchits;
